agents/pyproject.toml
[project]
name = "nexus-agents"
version = "0.1.0"
description = "NEXUS — Multiagent system powered by LangGraph"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2",
    "langchain>=0.3",
    "langchain-openai>=0.2",
    "langchain-anthropic>=0.2",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "psycopg2-binary>=2.9",
    "redis>=5.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "black>=24.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.black]
target-version = ["py311"]
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
